// src/app/interfaces/equipmentadditional.model.ts
export class Equipment
{
    item: any;
    quantity: any;
    ticketId: any;
    // version:number;
    // createdBy?:string;
    // type?:string;
    // status?:string;
    // priority?:string;
    // agencyId?:string;
    // description?:string;
    // ids?:string;
    // code?:string;
    // tech_assing?:string;
    // assigned_tags?:string;
    // assignedDate?: any;
    // closedDate?: any;
    // id?:number;
}

export class SerializaItems
{
    item_serial: any;
}